refactor: 기술부채 일괄 정리 (TX 경계/RoleHierarchy/Counter 통일/SHA-256/env) (#180)
* fix: LocalAuthService TX 경계 정상화

- register()의 createLocalAccount() self-invocation으로 @Transactional 무효화 해소
- TransactionTemplate으로 DB 쓰기 구간만 명시적 TX 적용
- authenticate()에서 불필요한 @Transactional(readOnly) 제거 (BCrypt/메트릭 TX 밖)
- createLocalAccount() 메서드 삭제

* refactor: DailyGameService TX 내 I/O 분리 + Counter 패턴 통일

- @Transactional 제거 후 TransactionTemplate으로 DB 쓰기만 TX 적용
- AI HTTP 호출/Redis 랭킹/이벤트 발행을 TX 밖으로 분리
- guessAnonymous()에서 불필요한 @Transactional(readOnly) 제거
- Counter 필드 @PostConstruct 등록으로 co-location 패턴 통일

* refactor: Counter 패턴 통일 (DiscordWebhookClient + CustomMetricsConfig)

- DiscordWebhookClient에 Counter 필드 @PostConstruct 등록으로 co-location
- CustomMetricsConfig의 businessCounters() MeterBinder 빈 삭제

* fix: SwaggerConfigController RoleHierarchy 적용

- SUPERADMIN이 admin Swagger 그룹에 접근하지 못하는 버그 수정
- RoleHierarchy.getReachableGrantedAuthorities()로 확장된 권한 확인

* refactor: BucketStore Clock 주입 + SHA-256 유틸 추출 + int cast 통일

- BucketStore의 Instant.now() 3곳을 clock.instant()로 교체
- HashUtils.sha256Hex() 유틸 추출, RefreshTokenService/TextNormalizer 위임
- AiServiceConfig의 (int) cast를 Math.toIntExact()로 통일

* chore: env 샘플 보완 (HINT_* 추가, RATE_LIMIT_READ 동기화)

- HINT_TRIGGER_THRESHOLD/HINT_MAX_SIMILARITY/HINT_MAX_COUNT 변수 추가
- RATE_LIMIT_READ를 application.yaml default(120)와 동기화

* style: React.* 네임스페이스 named import 전환

- RankingPanel.tsx: React.TransitionEvent -> import type { TransitionEvent }
- GuessInput.tsx: React.KeyboardEvent -> import type { KeyboardEvent }

* fix: DiscordWebhookClient Counter 초기화를 생성자로 이동

- @PostConstruct에서 생성자 초기화로 변경하여 단위 테스트 호환성 확보
- 테스트에서 수동 생성 시 @PostConstruct가 호출되지 않아 Counter NPE 발생하던 문제 해소

+++ backend/src/main/java/com/gakkaweo/backend/infra/ai/config/AiServiceConfig.java
@@ -10,7 +10,7 @@ public class AiServiceConfig {
 
   @Bean
   RestClient aiServiceRestClient(AiServiceProperties properties) {
-    int timeoutMillis = (int) properties.timeout().toMillis();
+    int timeoutMillis = Math.toIntExact(properties.timeout().toMillis());
 
     SimpleClientHttpRequestFactory factory = new SimpleClientHttpRequestFactory();
     factory.setConnectTimeout(timeoutMillis);
